AI/BT 구조 개선 및 대규모 NPC 동기화 최적화
BehaviorTree/AIComponent/서버 구조를 전면 개선하여, BT 기반 AI가 할당자와 블랙보드를 통해 안전하게 동작하도록 했습니다. NPC 업데이트 타이밍을 랜덤 분산(스태거링)하여 100마리 이상 동시 이동 시에도 부하가 분산되고, 패킷 폭주 현상이 해결됩니다. Room/Component/GameObject에 할당자 기반 Update를 추가하고, BT 노드 계층(Composite/Decorator/Leaf) 및 BTBuilder를 확장했습니다. 불필요한 타이머 기반 NPC 업데이트 코드를 제거하고, NPC 이동 패킷 배치 전송 임계값을 조정하는 등 서버 로직을 리팩터링했습니다. Lua/BT AI 공존이 가능하도록 AIComponent 구조도 유연하게 변경되었습니다.

+++ Server/Server/BT_Nodes.h
@@ -83,7 +83,7 @@ namespace PIP::GAME
                 lastPos = _blackboard->get<common::Vec3>("last_pos");
             }
 
-            float movedDist = std::sqrt(std::pow(current.x - lastPos.x, 2) + std::pow(current.z - lastPos.z, 2));
+            float movedDist = static_cast<float>(std::sqrt(std::pow(current.x - lastPos.x, 2) + std::pow(current.z - lastPos.z, 2)));
 
             // 예상 이동 거리의 20%도 못 갔으면 끼인 것으로 간주
             if (movedDist < (_speed * dt * 0.2f)) {
